Shifted test runner from test utils to run helper
